Put the mock back how it was before and update terraform to get the real redis url
