+++ lambda_functions/v1/functions/lpa/app/lpa_local.py
@@ -1,6 +1,6 @@
+from . import create_app
 from flask import Flask
 
-from .functions.lpa.app import create_app
 from .config import LocalMockConfig
 
 lambda_handler = create_app(flask=Flask, config=LocalMockConfig)
