+++ lambda_functions/v1/functions/lpa/app/lpa_mock.py
@@ -45,7 +45,7 @@ def rewrite_bad_request(response):
     return response
 
 
-mock = connexion.FlaskApp(__name__, specification_dir="openapi/")
+mock = connexion.FlaskApp(__name__, specification_dir="../../../openapi/")
 mock.app.after_request(rewrite_bad_request)
 mock.add_api("lpa-openapi.yml", strict_validation="true")
 mock.add_api("state-openapi.yml")
