+++ terraform/environment/lambda.tf
@@ -10,6 +10,7 @@ module "lambda_lpa_v1" {
   rest_api               = aws_api_gateway_rest_api.lpa
   account                = local.account
   redis_sg_id            = aws_security_group.lpa_redis_sg.id
+  redis_url              = aws_route53_record.lpa_redis.name
 }
 
 //Modify here for new version - create new one. keep original
